refactor: createUserInfo 순수 함수 분리로 테스트 시 Firebase 초기화 방지
- Extract createUserInfo to userInfoUtils.ts (pure function, no Firebase deps)
- stats.ts re-exports for backward compatibility
- commentingStatsUtils imports from userInfoUtils to break Firebase import chain
